add products srvice, dtocreateproduct, controller product with method createproduct, method getmpping for product actives, entitie product, relation with client in entities

+++ microservice-crm/src/main/java/com/microservice/crm/entities/Product.java
@@ -22,10 +22,10 @@ public class Product {
     private String name;
     
     @Column(precision = 6, scale = 2)
-    private BigDecimal pricePen;
+    private Double pricePen;
     
     @Column(precision = 6, scale = 2)
-    private BigDecimal priceDollar;
+    private Double priceDollar;
     
     @Column(nullable = false)
     private LocalDate start;
